move test to examples/transfers/transfer_test.go
